+++ tests/realtikvtest/importintotest/BUILD.bazel
@@ -2,8 +2,59 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
     name = "importintotest_test",
-    timeout = "short",
-    srcs = ["dummy_test.go"],
+    timeout = "long",
+    srcs = [
+        "detach_test.go",
+        "from_server_test.go",
+        "import_into_test.go",
+        "job_test.go",
+        "main_test.go",
+        "one_parquet_test.go",
+        "operate_test.go",
+        "precheck_test.go",
+        "show_test.go",
+        "util_test.go",
+    ],
+    embedsrcs = ["test.parquet"],
     flaky = True,
-    deps = ["//tests/realtikvtest"],
+    race = "on",
+    deps = [
+        "//br/pkg/lightning/backend/local",
+        "//br/pkg/lightning/common",
+        "//br/pkg/lightning/config",
+        "//br/pkg/mock",
+        "//br/pkg/mock/mocklocal",
+        "//br/pkg/streamhelper",
+        "//br/pkg/utils",
+        "//config",
+        "//disttask/framework/proto",
+        "//disttask/framework/scheduler",
+        "//disttask/framework/storage",
+        "//disttask/importinto",
+        "//domain/infosync",
+        "//executor",
+        "//executor/asyncloaddata",
+        "//executor/importer",
+        "//kv",
+        "//parser/auth",
+        "//parser/terror",
+        "//planner/core",
+        "//session",
+        "//testkit",
+        "//tests/realtikvtest",
+        "//util/dbterror/exeerrors",
+        "//util/sem",
+        "@com_github_docker_go_units//:go-units",
+        "@com_github_fsouza_fake_gcs_server//fakestorage",
+        "@com_github_golang_mock//gomock",
+        "@com_github_ngaut_pools//:pools",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_pingcap_kvproto//pkg/brpb",
+        "@com_github_pingcap_log//:log",
+        "@com_github_stretchr_testify//require",
+        "@com_github_stretchr_testify//suite",
+        "@io_etcd_go_etcd_client_v3//:client",
+        "@org_uber_go_atomic//:atomic",
+        "@org_uber_go_zap//:zap",
+    ],
 )
