--- tests/realtikvtest/importintotest/dummy_test.go
@@ -1,31 +1,0 @@
-// Copyright 2023 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package importintotest
-
-import (
-	"testing"
-
-	"github.com/pingcap/tidb/tests/realtikvtest"
-)
-
-func TestDummy(t *testing.T) {
-	realtikvtest.CreateMockStoreAndSetup(t)
-	// todo: delete it after enable importintotest.
-	t.Skip("skip dummy test")
-}
-
-func TestMain(m *testing.M) {
-	realtikvtest.RunTestMain(m)
-}
